move privateTransactionBuilder to own class
