+++ ethsigner/core/src/integration-test/java/tech/pegasys/ethsigner/jsonrpcproxy/model/jsonrpc/PrivateTransaction.java
@@ -91,72 +91,4 @@ public class PrivateTransaction {
   public String getRestriction() {
     return restriction;
   }
-
-  public static class PrivateTransactionBuilder {
-    private String from;
-    private String nonce;
-    private String gasPrice;
-    private String gas;
-    private String to;
-    private String value;
-    private String data;
-    private String privateFrom;
-    private List<String> privateFor;
-    private String restriction;
-
-    public PrivateTransactionBuilder withFrom(final String from) {
-      this.from = from;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withNonce(final String nonce) {
-      this.nonce = nonce;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withGasPrice(final String gasPrice) {
-      this.gasPrice = gasPrice;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withGas(final String gas) {
-      this.gas = gas;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withTo(final String to) {
-      this.to = to;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withValue(final String value) {
-      this.value = value;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withData(final String data) {
-      this.data = data;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withPrivateFrom(final String privateFrom) {
-      this.privateFrom = privateFrom;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withPrivateFor(final List<String> privateFor) {
-      this.privateFor = privateFor;
-      return this;
-    }
-
-    public PrivateTransactionBuilder withRestriction(final String restriction) {
-      this.restriction = restriction;
-      return this;
-    }
-
-    public PrivateTransaction build() {
-      return new PrivateTransaction(
-          from, nonce, gasPrice, gas, to, value, data, privateFrom, privateFor, restriction);
-    }
-  }
 }
